added check for passwords before updating user

+++ frontend/src/screens/ProfileScreen.js
@@ -37,29 +37,33 @@ export default function ProfileScreen() {
   const submitHandler = async e => {
     e.preventDefault();
 
-    try {
-      const { data } = await axios.put(
-        '/api/users/profile',
-        {
-          name,
-          email,
-          password
-        },
-        {
-          headers: { Authorization: `Bearer ${userInfo.token}` }
-        }
-      );
-      dispatch({
-        type: 'UPDATE_SUCCESS'
-      });
-      ctxDispatch({ type: 'USER_SIGNIN', payload: data });
-      localStorage.setItem('userInfo', JSON.stringify(data));
-      toast.success('User updated successfully');
-    } catch (err) {
-      dispatch({
-        type: 'FETCH_FAIL'
-      });
-      toast.error(getError(err));
+    if (password === confirmPassword) {
+      try {
+        const { data } = await axios.put(
+          '/api/users/profile',
+          {
+            name,
+            email,
+            password
+          },
+          {
+            headers: { Authorization: `Bearer ${userInfo.token}` }
+          }
+        );
+        dispatch({
+          type: 'UPDATE_SUCCESS'
+        });
+        ctxDispatch({ type: 'USER_SIGNIN', payload: data });
+        localStorage.setItem('userInfo', JSON.stringify(data));
+        toast.success('User updated successfully');
+      } catch (err) {
+        dispatch({
+          type: 'FETCH_FAIL'
+        });
+        toast.error(getError(err));
+      }
+    } else {
+      toast.error('Passwords do not match');
     }
   };
